+++ yash-syntax/src/parser/grouping.rs
@@ -24,6 +24,7 @@ use super::lex::Keyword::{CloseBrace, OpenBrace};
 use super::lex::Operator::{CloseParen, OpenParen};
 use super::lex::TokenId::{Operator, Token};
 use crate::syntax::CompoundCommand;
+use std::rc::Rc;
 
 impl Parser<'_, '_> {
     /// Parses a normal grouping.
@@ -85,7 +86,10 @@ impl Parser<'_, '_> {
             return Err(Error { cause, location });
         }
 
-        Ok(CompoundCommand::Subshell(list))
+        Ok(CompoundCommand::Subshell {
+            body: Rc::new(list),
+            location: open.word.location,
+        })
     }
 }
 
@@ -196,8 +200,12 @@ mod tests {
         let mut parser = Parser::new(&mut lexer, &aliases);
 
         let result = block_on(parser.compound_command()).unwrap().unwrap();
-        assert_matches!(result, CompoundCommand::Subshell(list) => {
-            assert_eq!(list.to_string(), ":");
+        assert_matches!(result, CompoundCommand::Subshell { body, location } => {
+            assert_eq!(body.to_string(), ":");
+            assert_eq!(*location.code.value.borrow(), "(:)");
+            assert_eq!(location.code.start_line_number.get(), 1);
+            assert_eq!(location.code.source, Source::Unknown);
+            assert_eq!(location.range, 0..1);
         });
     }
 
@@ -208,8 +216,12 @@ mod tests {
         let mut parser = Parser::new(&mut lexer, &aliases);
 
         let result = block_on(parser.compound_command()).unwrap().unwrap();
-        assert_matches!(result, CompoundCommand::Subshell(list) => {
-            assert_eq!(list.to_string(), "foo& bar");
+        assert_matches!(result, CompoundCommand::Subshell { body, location } => {
+            assert_eq!(body.to_string(), "foo& bar");
+            assert_eq!(*location.code.value.borrow(), "( foo& bar; )");
+            assert_eq!(location.code.start_line_number.get(), 1);
+            assert_eq!(location.code.source, Source::Unknown);
+            assert_eq!(location.range, 0..1);
         });
     }