+++ yash-syntax/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to `yash-syntax` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - Unreleased
+
+### Changed
+
+- `syntax::CompoundCommand::Subshell` from a tuple variant `Subshell(List)`
+  to a struct variant `Subshell { body: Rc<List>, location: Location }`.
+
 ## [0.5.0] - 2022-07-02
 
 This version contains variety of fixes.